temp: 임시코드 commit

+++ src/main/kotlin/calculator/Multiply.kt
@@ -1,6 +1,8 @@
 package calculator
 
+import java.math.BigDecimal
+
 class Multiply : Operation {
-    override fun calculate(firstOperand: Int, secondOperand: Int): Double =
-        firstOperand.toDouble() * secondOperand.toDouble()
+    override fun calculate(firstOperand: BigDecimal, secondOperand: BigDecimal): BigDecimal =
+        firstOperand * secondOperand
 }
